Build 4 - The Texture Derp Update
Added:
A new GUI, not perfect, but almost. Right click on any block to open
(this will be temp)
Blame my lazy texturers for making no textures, and in which actually
one of my coders made the dirt.
Made blocks smaller, from 28*28 to 24*24 (looks better)
Blocks are temp not tileable, until we get better and more permenent
textures